Add Recovery as Multisend (after deployment). (#20)
Adding recovery address as part of a multisend transaction (i.e. not included in deployment) whenever safe is not yet deployed and recovery address is provided. This ensures that the safeAddress linked to the near account is more deterministic. With this the Safe Address will be uniquely determined by nearAdapter.evmAddress and safeSaltNonce. Note that nearAdapter.evmAddress is also determined by a parameter similar to safeSaltNonce (but from Near to EVM). This additional parameter is called the derivationPath (usually defaulting to "ethereum,1").

Note that we also improve the balance check here (to ensure the safe is sufficiently funded in all scenarios).