Fix Bundler Sponsorship Policy (#121)
Previously this was filtering by this.chainId which was really returning
`policiesForChainId`.
Note that this method is chain agnostic although the the chainId and
entryPointAddress are required by the constructor. This hinted at
somewhat bad design that could be improved.

We split the Transaction Bundler from the Pimlico Service so that users
with a pimlico key can make those requests without having to specify all
the other irrelevant details.

Tests Added.